feat/re-allow multiple workers (#36134)
#### Description

This MR does the following:

* Re adds the ability to allow multiple workers in this exporter due to:

1. Out of Order is no longer an issue now that it is fully supported in
Prometheus. Nonetheless, I am setting the default worker as 1 to avoid
OoO in Vanilla Prometheus Settings.

2. With a single worker, and for a collector with a large load, this
becomes "blocking". Example: Imagine a scenario in which a collector is
collecting lots of targets, and with a slow prometheus/unstable network,
a single worker can easily bottleneck the off-shipping if retries are
enabled.

#### Link to tracking issue
N/A

#### Testing  ####

#### Documentation
docs auto-updated. Readme.md is now correct in its explanation of the
`num_consumers since its no longer hard-coded at 1. Additional docs
added.