[exporter/loadbalancing] Fix shutdown race (#37477)
Fix #37446 - The scope of the `sync.WaitGroup` used to ensure that a
periodic function called from a background goroutine is finished is
incorrect, allowing the function to be started after the shutdown was
"completed". The fix just moves the calls to `Add` and `Done` to the
correct scope.

Although a bug on the component it doesn't seem to deserve a changelog
since it was detected via tests and there is no issue around the
shutdown of the component.