[connector/datadog] support obfuscation in datadogconnector (#37457)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Support obfuscating SQL and Redis queries in APM stats in Datadog
Connector.

replace statement in otelcontribcol was necessary due to failing build
in pipeline. It doesn't appear that any modules use any affected
functions, but building otelcontribcol with ocb failed due to an API
change. Also opened an issue to bump the `receiver/awscontainerinsight`
dependency so that this replace can be removed in the future: #37486

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Replaces #35401

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added TestObfuscate in connector_native_test.go

<!--Describe the documentation added.-->
#### Documentation
added release note and link to semantics on Datadog vendor website
<!--Please delete paragraphs that you did not use before submitting.-->