[chore] Generate `replaces` automatically in otelcontribcol/oteltestbedcol (#37569)
#### Description
`cmd/otelcontribcol/builder-config.yaml` currently contains numerous
`replaces` directives. These are important for making sure that builds
use the local version of the code and not one fetched from the Go proxy,
especially during the release process when internal dependencies are set
to versions that have not been published yet. To avoid needing to
maintain this list of directives manually, this PR removes them from
`builder-config.yaml` and adds a script which creates
`builder-config-replaced.yaml` by appending `replaces` directives for
every module in the repository. (Some may be superfluous, but that
should not cause problems.)

#### Link to tracking issue
Fixes #37243

#### Testing
I ran the release script locally and no errors occurred in
`genotelcontribcol` or `gentestbedcol`.