[exporter/elasticsearch] use v8 client, drop support for EOL versions of ES (#37333)
#### Description

Update to go-elasticsearch v8, which breaks support for older versions
of Elasticsearch v7. The non-EOL v7.17.x versions of Elasticsearch are
still supported by the exporter. Add an explicit statement to the README
that EOL versions are not supported by the exporter.

Updating to the v8 client allows us to turn on OTel tracing
instrumentation for the go-elasticsearch client.

#### Link to tracking issue

Closes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/32454

#### Testing

- Tested with 8.latest, confirmed working
- Tested with 7.17.lasted, confirmed working
- Tested with 7.10 (EOL), confirmed that it fails with an error message:

> 2025-02-05T13:23:42.788+0800 error
elasticsearchexporter/bulkindexer.go:337 bulk indexer flush error
{"kind": "exporter", "data_type": "metrics", "name": "elasticsearch",
"error": "failed to execute the request: the client noticed that the
server is not Elasticsearch and we do not support this unknown product"}

(7.10 predates the product check protocol that was added to
go-elasticsearch v8.)

#### Documentation

Updated README.