[processor/resourcedetection] Add Dynatrace resource detector (#37766)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR adds the Dynatrace detector to the resource detection processor.
The Dynatrace detector loads resource information from the
`dt_host_metadata.properties` file which is located in
the `/var/lib/dynatrace/enrichment` (on *nix systems) or
`%ProgramData%\dynatrace\enrichment` (on Windows) directories.
If present in the file, the following attributes will be added:

- `dt.entity.host`
- `host.name`

The Dynatrace detector does not require any additional configuration,
other than being added to the list of detectors.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #37577

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added unit tests

<!--Describe the documentation added.-->
#### Documentation
Added a description of the detector to the readme

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Signed-off-by: Florian Bacher <florian.bacher@dynatrace.com>
Co-authored-by: Evan Bradley <11745660+evan-bradley@users.noreply.github.com>