[receiver/syslog] Support setting on_error mode for syslog receiver (#37847)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
As per the discussion in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36906,
this PR supports setting the `on_error` mode in syslog receiver that
controls the behaviour of the underlying `syslog_parser` when it
encounters an error.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Closes
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36906.

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Tested setting `on_error` in the syslog receiver config.

<!--Describe the documentation added.-->
#### Documentation
Documented the `on_error` entry in both the syslog input and receiver
config.
<!--Please delete paragraphs that you did not use before submitting.-->

Signed-off-by: Mengnan Gong <namco1992@gmail.com>