[exporter/stefexporter] Fix a context cancellation bug in STEF exporter (#37944)
#### Description

STEF exporter used context that was passed to exportMetrics() as the
context for the entire gRPC stream.

This was wrong since the context is cancelled immediately after
exportMetrics returns.

The exporter now uses a separate context for the gRPC stream. This
context has a longer duration, matching the duration of the gRPC stream.

The context passed to exportMetrics() continues to be used as the
context for connection attempt. If connection is not established within
the limits of that context then the attempt to connect is correctly
aborted.

This now correctly decouples the context (and cancellation) for
connection attempt from the context of the connected gRPC stream.

#### Testing

Added unit tests to verify the bug fix.