fix(azuremonitorreceiver): Azure Monitor receiver should not produce gaps in data points for PT1M time grains (#37342)
#### Description
The current implementation of tracking for the last metrics fetch is
highly sensitive to time, relying on `time.Now()`. This sensitivity
causes the loss of data points due to jitter in 1-minute intervals. I
recommend a change request to implement a more robust comparison with
the last minute of metrics fetch.

#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/37337

#### Testing
A new coverage introduced in unit test

#### Documentation
No documentation added

---------

Co-authored-by: Curtis Robert <crobert@splunk.com>