[receiver/kafka_metrics_receiver] Add refresh_frequency to enable custom refreshing for cluster metadata (#37897)
Fixes:
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/37896

There are other ways to solve this issue, but that would require changes
in Sarama library.
This PR proposes a solution to exposes `refresh_frequency` from Sarama
library and let the user configure it.

Today the default setting from Sarama is being used which defaults to 10
mins.

More info and similar issues are mentioned in the tracking issue.

Added unit tests, and tested manually with a 3 brokers kafka-kraft
cluster.

---------

Signed-off-by: Shivanshu Raj Shrivastava <shivanshu1333@gmail.com>
Co-authored-by: Dmitrii Anoshin <anoshindx@gmail.com>