[chore] File issue when update-otel workflow fails (#38032)
#### Description
This PR records the output of the `update-otel` workflow into a file,
and creates an issue with 'ci-cd' label containing the last 100 lines
and a link to the full log if the main step fails.

#### Link to tracking issue
Resolves #37679

#### Testing
I tested the code on my fork, with a dummy job that always fails
([workflow
logs](https://github.com/jade-guiton-dd/opentelemetry-collector-contrib/actions/runs/13396711147),
[resulting
issue](https://github.com/jade-guiton-dd/opentelemetry-collector-contrib/issues/8)).