[telemetrygen] support `--service` for all signal types (#38044)
This flag was supported only for traces. It's possible to set the
service name via `--otlp-attributes`, but it's more convenient to be
able to set it via `--service`. I added it to all signals, but if others
disagree with this change, then I would suggest we remove it from all
signals.

---------

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>