[chore] [pkg/stanza] test: add benchmark for File input (#38054)
This benchmark has a scope that sits between the [Stanza File consumer
benchmark](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.120.0/pkg/stanza/fileconsumer/benchmark_test.go)
and a File Log receiver benchmark (which currently does not exist).

The major difference between the File consumer benchmark and the File
input benchmark is that the File input benchmark includes measuring of
the memory allocations made in the File input's
[emit](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.120.0/pkg/stanza/operator/input/file/input.go#L40)
function.

This should allow to assess performance impact of this change:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/37734,
or any similar changes in the future.