[connector/signaltometrics]Drop get OTTL func in favor of OTTL's ParseValueExpression (#38101)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

The component was using an internal `get` OTTL function as a stop gap
measure to extract value from OTTL expressions due to lack of such
function from the core OTTL library. With
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/35621,
support for data extraction using OTTL value expressions was introduced.
This PR uses the `ParseValueExpression` method and drops the `get` OTTL
function. The function was documented in README as internal function and
strictly advised not to be used in configurations so I don't expect any
issues.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/38098

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Unit tests are passing

<!--Describe the documentation added.-->
#### Documentation

README is updated

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Andrzej Stencel <andrzej.stencel@elastic.co>