[opampsupervisor] add telemetry settings to the opamp supervisor (#38017)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR adds options to configure the telemetry settings for the OpAmp
supervisor. This will enable the supervisor, in addition to logging,
export metrics and traces as well. For the sake of consistency, I also
moved the logger of the supervisor into the telemetrySettings property
of the supervisor.
This PR does not add any traces or metrics yet, but just provides the
groundwork to add those later on.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #35582

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added Unit tests. E2E test can be added once we start emitting traces
and metrics.

---------

Signed-off-by: Florian Bacher <florian.bacher@dynatrace.com>
Co-authored-by: Evan Bradley <11745660+evan-bradley@users.noreply.github.com>