[extension/encoding/awscloudwatchmetricstreams_encoding] Support unmarshalling JSON-formatted CloudWatch Metric Streams (#38419)
#### Description

This PR implements the translation of JSON-formatted CloudWatch Metric
Streams into OpenTelemetry metrics.

It receives a record, and splits that record by each new line. Each of
these parts will correspond to a JSON CloudWatch Metric Stream. It then
translates this metric into an OpenTelemetry metric. The obtained
metrics from the record are returned.

It follows the same logic as
https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/receiver/awsfirehosereceiver/internal/unmarshaler/cwmetricstream.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/38407.

<!--Describe what testing was performed and which tests were added.-->
#### Testing

I have added unit tests to:

- Check we can extract open telemetry metrics from JSON valid cloudwatch
metric stream
- Check there are no metrics extracted if the data does not have
expected format (for example, no unit or no value)
- Check we can obtain metrics from a record containing both valid and
invalid cloudwatch metric stream.

You can easily read and study what we get to what we return in the
`testdata/json` files of the extension.

<!--Describe the documentation added.-->
#### Documentation

The coding comments and the README file should be enough for
documentation.

---------

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>
Co-authored-by: Sean Marciniak <30928402+MovieStoreGuy@users.noreply.github.com>