[googlecloud owners] chore: update gcp owners (#38617)
We remove punya from the owners list and add in Dylan and Ridwan.

<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
We are adding new members of the team at GCP as maintainers of the GCP
components here and taking one member away.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
N/A

<!--Describe what testing was performed and which tests were added.-->
#### Testing
N/A

<!--Describe the documentation added.-->
#### Documentation
The change is updating documentation and metadata.
<!--Please delete paragraphs that you did not use before submitting.-->