[chore][ping-code-owners-on-new-issue] Fix variable reference (#38840)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Lately some `Ping code owners on a new issue` action runs have [been
failing](https://github.com/open-telemetry/opentelemetry-collector-contrib/actions/runs/13972340605/job/39117145364)
with the following error:
```
./.github/workflows/scripts/ping-codeowners-on-new-issue.sh: line 36: COMPONENT: unbound variable
```

This was because recently a variable reference was changed to the wrong
variable name, view the diff of the related PR to see the context.

Related:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38622