[extension/azureauthextension] Use credentials to token (#39057)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

The previous implementation required a scope to be defined, and handled
all the logic related to the access token (updating it, caching it,
etc). With this change, only the credentials are now responsible to get
the access token, and we completely rely on the azure SDK.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
N/A

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Unit tests fixed.

<!--Describe the documentation added.-->
#### Documentation

README is updated.

<!--Please delete paragraphs that you did not use before submitting.-->