chore: update mongodb driver library to use v2 (#39268)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #39257

Major version upgrade got rid of the mtest package, there was some
guidance here but the strategy they recommend is listed as deprecated
https://www.reddit.com/r/golang/comments/1fvb9d7/how_do_you_write_unit_test_for_mongogodriver_v2/

I've added some golangci-lint skips to accomodate for this but should be
noted for future upgrades that it may require a different approach to
update the general scraper tests. We are already employing containerized
mongo for testing using our integration tests so we should feel
confident if those continue to pass on our support matrix.

With the driver update a lot of the docs returned from mongo are now
`bson.D` which required some code updates to our dig function which is
included in this PR.


#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Antoine Toulme <atoulme@splunk.com>