[opampsupervisor] Allow the agents to start be started without receiving a remote config (#39476)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR allows the Supervisor to start agents without having received a
remote configuration from the backend, which is a totally valid scenario
after the addition of the local configuration feature (see #38671).

This is achieved by setting `configMapIsEmpty` within
`Supervisor.composeMergedConfig` only IF the remote config msg given to
it is not `nil`. It is necessary because this method is not only called
when handling remote config msgs (which guarantees they are not nil),
but also when setting up "own telemetry" and processing agent
identification. In these two places it is given
`Supervisor.remoteConfig` -- potentially `nil` if no remote config has
been received or the server has no `OffersRemoteConfig` capability.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #38794. 

<!--Describe what testing was performed and which tests were added.-->
#### Testing

- Added unit tests covering more cases of the
`Supervisor.composeMergedConfig` method and changed the test itself to
use table tests.
- Added an e2e test covering the case where a Supervisor starts without
last received config, with nop backend, and (valid) local configuration.

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->