kafkareceiver: refactor to make consumer generic (#39360)
#### Description

Remove code that is repeated for every signal, and implement a generic
consumer that calls out to a signal-specific message handler.

We also update all the tests to use kfake rather than mocks so we're
exercising all of the receiver. These tests are parallelised, so all
tests run faster than before, and with greater coverage.

Finally, we add `max_fetch_wait` (defaults to 250ms) to the receiver's
configuration. This was added for speeding up the tests, but may also be
useful for users of the receiver.

#### Link to tracking issue

N/A

#### Testing

See description - tests now use the real client, rather than mocks,
using https://github.com/twmb/franz-go/tree/master/pkg/kfake.

#### Documentation

Updated README.

---------

Co-authored-by: Sean Marciniak <30928402+MovieStoreGuy@users.noreply.github.com>