elasticsearchexporter: mapping mode from scope (#39110)
#### Description

Add support for specifying the mapping mode as a scope attribute,
"elastic.mapping.mode". If specified, this overrides both
X-Elastic-Mapping-Mode from client metadata as well as the default mode
specified through collector configuration.

Simplify StartSession by removing the error result, since the only time
it can fail is if the BulkIndexerConfig is invalid, and we ensure that
cannot happen. We now also pre-create arrays of encoders, routers, and
data point hashers; this simplifies and speeds up acquisition of these
during execution.

This is based on
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38667

#### Link to tracking issue

N/A

#### Testing

Added unit tests.

#### Documentation

Updated README.