[connector/signaltometrics] Fix metric identity to use name, type, and unit (#39454)
#### Description

The component used metric name and description to identify metric that
should be produced. Due to this, if duplicate metric names with
different metric types were added then it would lead to incorrect
results. The PR fixes this case by using the correct metric identity.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/39442

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added testcase for logs and traces

<!--Describe the documentation added.-->
#### Documentation

N/A

---------

Co-authored-by: Christos Markou <chrismarkou92@gmail.com>