[prometheusreceiver]: allow mixed classic and native histogram (#39451)
Depends on #38645 for testing.

#### Description

This is technical dept from #28663, especially in migration
scenarios we want to receive both versions of the same histogram,
the classic explicit and exponential bucket histogram.

The current internal model of the receiver does not allow this, so
I modified the model to not only key metrics by metric family name,
but also whether it's an exponential histogram.

Some callbacks from Promtheus scrape don't tell us what the metric type
is up front.

For AppendExemplar to work I made transaction stateful to remember if it
appended an exponential histogram before.
For Append of staleness marker to work, I modified the algorithm to
discover exponential histogram from the cached metadata and the name.
Due to https://github.com/prometheus/prometheus/issues/16217 I also had
to modify AppendExemplar to detect trying to add exemplars from native
histograms even if native histograms are turned off.

#### Link to tracking issue
Related to #26555 

#### Testing
Unit test updated.

#### Documentation
N/A - does not contradict.

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>