Fail workflow when conclusion is failure (#39805)
#### Description
In
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38270,
I've introduced a workflow to create new issues when tests fail on main.
Somehow, that change caused an unexpected behavior with test result
reports, as explained in
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/39616.

To be completely transparent, I'm unsure if this PR fixes the issue.
Following what is [documented by
GitHub](https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/accessing-contextual-information-about-workflow-runs#steps-context),
the existing configuration should work... but, well, it isn't. So I'm
trying the opposite of what the documentation says.

I would really appreciate it if someone with GitHub Actions experience
could chime in and guide me because I'm taking shots in the dark at this
point 😕

Signed-off-by: Arthur Silva Sens <arthursens2005@gmail.com>
Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>