Support named encoding extensions in Kafka receiver and exporter (#40144)
#### Description

Fix the Kafka receiver and exporter to support encoding extensions with
a custom name, e.g. "otlp_encoding/json" rather than assuming it is only
a type, e.g. "otlp_encoding".

Note: in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/39809/files#r2069988901
it was suggested that we use `component.ID` in the config struct. That
won't work for the Kafka receiver since it also has built-in encodings
that are invalid component IDs. It could work for the exporter, but I
have opted to keep the code consistent instead.

#### Link to tracking issue

Fixes #40142

#### Testing

Added tests.

#### Documentation

None