[chore] apply unkeyed literal initialization prevention (#39832)
Adds a new check on checkapi that prevents the unkeyed initialization of
literals. Unkeyed initialization breaks if the API exposed by the struct
fields changes. Adding an anonymous _ struct{} field ensures all
developers must use a named initialization.

This is explained in more detail in
https://github.com/open-telemetry/opentelemetry-collector/issues/12360

This PR needs #39830 and #39831 merged first.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>