[pkg/experimentalmetricmetadata] Add missing otel.entity.type fields to delete events (#40279)
Add missing `otel.entity.type` field to the delete events. The type
along with the set of identifying attributes forms an entity identity.
So it has to be present in all types of entity events. This change adds
new fields to the `EntityDeleteDetails`, but the entity type fields will
be moved to the event level in the next PR to avoid redundancy in the
API