fix(sumologicextension): process discovery (#40253)
#### Description
This change makes the sumologicextension more tolerant of process
discovery failures. Previously, any error in getting a process name
could cause the sumologicextension to fail to connect. This failure can
happen in the case of zombie processes, or for other reasons, as these
calls involve system calls.

In 2024, an exception was made so that Windows only reported warnings
when this happened, as Windows seemed to be more susceptible to this
happening. However, we now have evidence that it can happen on other OS
as well, so it's best to warn for other OS too.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #40252

<!--Describe what testing was performed and which tests were added.-->
#### Testing
It's difficult to test this scenario, we will be manually verifying the
change where breakage was observed in development. gopsutil does not
seem to include interfaces that we can mock, and the errors are
ultimately emanating from syscalls.

---------

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>