[chore] Fix config example (#40627)
This PR fixes the config example which might be misleading. 

The existing example includes a `k8sattributes` processor instance which
is empty (fall backs to default) and a one `k8sattributes/2` with
actually configuration. I think we can remove the "empty" one since it
doesn't serve much as an example.