[receiver/simpleprometheus] Pass labels configuration to Prometheus properly (#40723)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Provides fix as described in the issue. The original issue was the the
receiver was not properly configuring the Prometheus receiver, which
this is a wrapper for. This change is to simply pass the labels under
the `Labels` Prometheus receiver option, rather than under `Target`.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #40722

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Manually tested on CLI. There aren't good tests for this component,
shown by this bug not being caught earlier. If anyone has any ideas for
good ways to add tests here, I'd be happy to add them.

<!--Please delete paragraphs that you did not use before submitting.-->