Fix datatype discrepancies in Snowflake Receiver (#40777)
#### Description
Some of the metrics of snowflake receiver didn't have the correct
datatype. Details
[here](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/40418#issuecomment-2980322405).
This PR fixes the discrepancies and improves the unit testing.


#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/40418

#### Testing

The current tests were testing just the type of the metrics.
Added the testing of values to the unit tests. 

Changed the value of bytesInserted to Float in the unit test.

#### Documentation

Added the changelog entry