[receiver/windowsservice] Added receiver scrape wireframe (#40043)
#### Description
This PR adds the outline for the scraping functionality to the Windows
Service receiver. This outlines the intended structure including how the
receiver will access the win32 api and serialize the results. The actual
implementation details will be submitted in a separate PR once the
approach and framework is agreed upon.

The primary motivation for breaking the PRs up like this is to prevent
scope creep and keep code light and easy to review.

#### Link to tracking issue
Fixes
[38619](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/38619)

#### Testing
No testing required at this point.

#### Documentation
README.md included as well as scratch notes on the components.