[connector/datadog] - Use waitgroup instead of closing a channel and waiting on it (#40849)
#### Description

The `run()` goroutine closes the channel in defer. 
There are two sceanrios:
1. Best case scenario: no error encountered while calling
`ConsumeMetrics`.
2. Worst case scenario: we encounter an error while calling
`ConsumeMetrics`, causing the goroutine to exit early and as a result,
close the `c.exit` chan.

We close the channel, again, in `Shutdown` and this results in a "send
on closed channel" panic in worst case scenario. In best case scenario,
no panic is encountered.

This PR adds waitgroup to avoid the panic.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #40845

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added a test case