[receiver/prometheus] Cleanup and restructure README (#41251)
#### Description

The README of this receiver is really hard to read. It seems like, over
time, new sections have been inserted at really bad places, disrupting
the flow of existing text.

For example, just below the `Feature gates` section there was this text:

```
- `report_extra_scrape_metrics`: Extra Prometheus scrape metrics can be reported by setting this parameter to `true`

You can copy and paste that same configuration under:

``yaml
receivers:
  prometheus:
    config:
``
```

* What is `report_extra_scrape_metrics` talking about? What does this
have to do with the `RemoveStartTimeAdjustment` feature gate?
* What is `You can copy and paste that same configuration under:`
talking about? What "configuration"? The Feature gates?

It took a while to figure out what the README tries to tell me.

I've moved the "Feature Gates" paragraph as its own section to the end
of the file, to be more in line with the other READMEs.