[exporter/loadbalancing] Use Linear probing to deal with hash collisions, and increase uniformity of endpoint distribution (#41271)
#### Description

Updates the ring hash algorithm to use linear probing to decrease the
variance of endpoints chosen due to hash collisions.
 
Also updates it to use 4 bytes instead of one - this would cause
collision when using more than 256 endpoints, since the byte would
overflow.

These changes are discussed and benchmarked in:
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41200.
The benchmarks showed clear benefits for negligible resource
consumption.

The only actual changes to main code in the PR is that in
`consistent_hashing.go`. The rest is just changing the expectations of
the test to match the altered hashing algorithm.

#### Link to tracking issue

Partially fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41200

#### Testing

The tests needed to be re-arranged. This was really quite tedious, since
they all depend on the specific randomness of the existing hashing
algorithm. So I had to re arrange some of the test's expectations with
that, too. At some point, they should be redesigned to be more
deterministic so more work can easily be done on the hash ring in the
future.