[receiver/awscloudwatch]: fix issue with polling of deleted log group while polling (#40571)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Resolves #38940

If a log group was deleted while a poll was being executed we were
silently failing after the first log message. While examining the code I
don't believe we were failing on future polls but can imagine that this
change is an improvement as it now radiates a tad bit more on the
`pollForLogs` function as it now actually returns up errors.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #38940


<!--Describe what testing was performed and which tests were added.-->
#### Testing

Unit testing the situation described in the issue

#### Documentation

changelog entry