Enrich telemetry from client metadata as experimental feature (#41284)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Client metadata carries important information that is sometimes
essential as a metric dimension for internal telemetry.
https://github.com/open-telemetry/opentelemetry-collector/issues/4970 is
trying to address the issue on a collector level - this PR aims to
provides a short-term, experimental, solution to this problem.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
N/A

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests added

<!--Describe the documentation added.-->
#### Documentation
N/A

<!--Please delete paragraphs that you did not use before submitting.-->