[extension/filestorage] - Add an option to recover from panic if underlying file is corrupted. (#40986)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Add an option to recreate the database from scratch. This is useful if
the underlying database is corrupted and there's no way to recover from
it.

Ideally, we would want to `recover` from a panic. However, I came across
the `panic` statements in bbolt's codebase and they are raised inside of
a goroutine. Unfortunately, we [can't
recover](https://go.dev/wiki/PanicAndRecover#:~:text=A%20panic%20cannot%20be%20recovered%20by%20a%20different%20goroutine)
from that panic in our collector process. So, our best route here is to
provide an option to the user to recreate a new database.


<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #36840, #35899

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Mikołaj Świątek <mail@mikolajswiatek.com>