[receiver/awscloudwatch]: fix ensure close on storage clients opened by the awscloudwatchreceiver (#41261)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Was noticing while using this exporter sometimes with rapid restarts
while using the `file_storage` extension that we get a timeout trying to
access the extension again. I believe this is improper usage of the
storage adapter and we should close it on shutdown on the receiver
component as well. Not entirely sure if its the extension's
responsibility for this but

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #41215

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->