[exporter/tinybird] implement metrics propagation (#41364)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Implement metrics propagation for the new Tinybird Exporter. The
exporter iterates over the pmetric data, extracts the required fields
(service name, attributes, spanID, etc.), generates an NDJSON, and
performs a request to the Tinybird
[EventsAPI](https://www.tinybird.co/docs/forward/get-data-in/events-api)
with all the data.

It's the same implementation done in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/40993
and
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/41248,
but this time focused on metrics. The only difference is that instead of
just having a `SignalConfig` for `Metrics`, it has one for each type of
metric (`MetricsGauge`, `MetricsSum`, `MetricsHistogram`, and
`MetricsExponentialHistogram`). The reason is that each type of metric
is stored in a different datasource in Tinybird (each has different
schema). This is based on the clickhouseexporter.

Now that the three signals have been implemented, the component
lifecycle tests have been re-enabled

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Related to
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/40475

<!--Describe the documentation added.-->
#### Documentation
- Modified the exporter README to add the per metric configuration

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: MoreraAlejandro <amorera@tinybird.co>