[exporter/tinybird] default logs' Timestamp value to ObservedTimestamp when Timestamp is not set (#41447)
#### Description

In some cases (like using Rust OTel SDK), the Timestamp value for the
logs signal was not correctly set. That caused log entries to be set to
the start timestamp epoch, which is not the desired behavior. For that
reason, we are defaulting the Timestamp value to the ObservedTimestamp
value, so we ensure there is always a valid value set.

#### Link to tracking issue
Fixes #41397