[receiver/k8sobjectsreceiver] Implement include_initial_state for watch mode (#41564)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This PR adds a global `include_initial_state` option to the receiver's
watch mode to enable capturing the existing state of objects as logs,
before the receiver starts watching them in real time.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41536

<!--Describe what testing was performed and which tests were added.-->
#### Testing

- `TestCreateDefaultConfigIncludeInitialState`
- `TestConfigValidationIncludeInitialState`
- `TestIncludeInitialState`
- `TestIncludeInitialStateWithPullMode`

<!--Describe the documentation added.-->
#### Documentation

I've updated the configuration section in the README and the config
example with these changes.