[receiver/httpcheck] Add optional request body (#41381)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Adds a body configuration that will be sent to the HTTP check endpoint
to allow testing specific actions.

The default is the current empty body if it is not configured.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #41325 

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Adds unit tests to check that these metrics are behaving as expected.

<!--Describe the documentation added.-->
#### Documentation

The readme shows how to enable this capability in the configuraiton.

<!--Please delete paragraphs that you did not use before submitting.-->